rearrages deb files
deb files were prefixed by build which caused them to end up getting
installed into build/*, in the same directory as the deb package
instead of the intended path (i.e., usr/*).

this fixes the problem by getting fpm to change dir to build, and
instructing it to build the usr dir.