Update cs50.c
Since get_string acts like a parser for the other functions the "const" or in this case "const char* " return type explicitly tells the student that get_string reads the user's input and that result cannot be changed(by accident). It secures that the input cannot be changed along the way.