cautious-1
Since the design paradigm in CS50.c is such that static variables and pointers are explicitly assigned to 0 and NULL respectively, I think it's more important for not so obvious things to be pointed out, such as string is. I backup my standings by the confusion it may case with the difference between local variables which if not defined will hold some "garbage" value and pointers such those ones which are set to point to NULL by default even without being static by nature. If students are well familiar with this property, I think they'll pay more attention when assert-ing/checking later on. I'll be honest enough to admit that when learning about it, I was mislead to think that something like "garbage" memory could exist. :)